+++ packages/vue/framework7-vue.js
@@ -1,5 +1,5 @@
 /**
- * Framework7 Vue 3.0.0-beta.8
+ * Framework7 Vue 3.0.0-beta.9
  * Build full featured iOS & Android apps using Framework7 & Vue
  * http://framework7.io/vue/
  *
@@ -7,7 +7,7 @@
  *
  * Released under the MIT License
  *
- * Released on: June 11, 2018
+ * Released on: June 12, 2018
  */
 
 (function (global, factory) {
@@ -1440,7 +1440,8 @@
       raised: Boolean,
       outline: Boolean,
       active: Boolean,
-      disabled: Boolean
+      disabled: Boolean,
+      tooltip: String
     }, Mixins.colorProps, Mixins.linkIconProps, Mixins.linkRouterProps, Mixins.linkActionsProps),
 
     render: function render() {
@@ -1488,6 +1489,7 @@
       }
 
       return _h('a', __vueComponentTransformJSXProps(Object.assign({
+        ref: 'el',
         style: style,
         class: self.classes
       }, self.attrs, {
@@ -1581,7 +1583,31 @@
         __vueComponentDispatchEvent.apply(void 0, [ this, events ].concat( args ));
       }
 
+    },
+
+    mounted: function mounted() {
+      var self = this;
+      var ref = self.props;
+      var tooltip = ref.tooltip;
+      if (!tooltip) { return; }
+      self.$f7ready(function (f7) {
+        self.f7Tooltip = f7.tooltip.create({
+          el: self.$refs.el,
+          text: tooltip
+        });
+      });
+    },
+
+    beforeDestroy: function beforeDestroy() {
+      var self = this;
+
+      if (self.f7Tooltip && self.f7Tooltip.destroy) {
+        self.f7Tooltip.destroy();
+        self.f7Tooltip = null;
+        delete self.f7Tooltip;
+      }
     }
+
   };
 
   var F7CardContent = {
@@ -2895,7 +2921,8 @@
       href: {
         type: [String, Boolean],
         default: '#'
-      }
+      },
+      tooltip: String
     }, Mixins.colorProps, Mixins.linkIconProps, Mixins.linkRouterProps, Mixins.linkActionsProps),
 
     data: function data() {
@@ -3004,6 +3031,7 @@
       var ref = self.props;
       var tabbarLabel = ref.tabbarLabel;
       var tabLink = ref.tabLink;
+      var tooltip = ref.tooltip;
       var isTabbarLabel = false;
 
       if (tabbarLabel || (tabLink || tabLink === '') && self.$$(el).parents('.tabbar-labels').length) {
@@ -3013,6 +3041,23 @@
       self.setState({
         isTabbarLabel: isTabbarLabel
       });
+      if (!tooltip) { return; }
+      self.$f7ready(function (f7) {
+        self.f7Tooltip = f7.tooltip.create({
+          el: self.$refs.el,
+          text: tooltip
+        });
+      });
+    },
+
+    beforeDestroy: function beforeDestroy() {
+      var self = this;
+
+      if (self.f7Tooltip && self.f7Tooltip.destroy) {
+        self.f7Tooltip.destroy();
+        self.f7Tooltip = null;
+        delete self.f7Tooltip;
+      }
     },
 
     computed: {
@@ -9037,7 +9082,13 @@
         var pageData = {
           component: component,
           id: id,
-          props: Utils.extend({}, options.route.params),
+          props: Utils.extend(
+            {
+              f7route: options.route,
+              f7router: router,
+            },
+            options.route.params
+          ),
         };
         routerComponent.$f7router = router;
         routerComponent.$f7route = options.route;
@@ -9110,7 +9161,13 @@
         var tabContent = {
           id: id,
           component: component,
-          props: Utils.extend({}, options.route.params),
+          props: Utils.extend(
+            {
+              f7route: options.route,
+              f7router: router,
+            },
+            options.route.params
+          ),
         };
 
         tabsComponent.$f7router = router;
@@ -9161,7 +9218,13 @@
         var modalData = {
           component: component,
           id: id,
-          props: Utils.extend({}, options.route.params),
+          props: Utils.extend(
+            {
+              f7route: options.route,
+              f7router: router,
+            },
+            options.route.params
+          ),
         };
         modalsComponent.$f7router = router;
         modalsComponent.$f7route = options.route;
@@ -9199,7 +9262,7 @@
   };
 
   /**
-   * Framework7 Vue 3.0.0-beta.8
+   * Framework7 Vue 3.0.0-beta.9
    * Build full featured iOS & Android apps using Framework7 & Vue
    * http://framework7.io/vue/
    *
@@ -9207,7 +9270,7 @@
    *
    * Released under the MIT License
    *
-   * Released on: June 11, 2018
+   * Released on: June 12, 2018
    */
 
   var Plugin = {
@@ -9340,9 +9403,11 @@
       Object.defineProperty(Extend.prototype, '$f7route', {
         get: function get() {
           var self = this;
+          if (self.props && self.props.f7route) { return self.props.f7route; }
+          if (self.f7route) { return self.f7route; }
+          if (self._f7route) { return self._f7route; }
           var route;
           var parent = self;
-          if (self._f7route) { route = self._f7route; }
           while (parent && !route) {
             if (parent._f7route) { route = parent._f7route; }
             {
@@ -9359,9 +9424,11 @@
       Object.defineProperty(Extend.prototype, '$f7router', {
         get: function get() {
           var self = this;
+          if (self.props && self.props.f7router) { return self.props.f7router; }
+          if (self.f7router) { return self.f7router; }
+          if (self._f7router) { return self._f7router; }
           var router;
           var parent = self;
-          if (self._f7router) { router = self._f7router; }
           while (parent && !router) {
             if (parent._f7router) { router = parent._f7router; }
             else if (parent.f7View) {